+++ .gitignore
@@ -3,4 +3,6 @@ target/
 .vscode/
 build
 build_*
+*.so
+*.dll
