+++ src/main/java/by/bcrypto/bee2j/provider/Bee2SecurityProvider.java
@@ -39,13 +39,13 @@ public final class Bee2SecurityProvider extends Provider {
         put("Signature." + JceNameConstants.BignWithBash384 + " SupportedKeyClasses", bignKeyClasses);
         put("Signature." + JceNameConstants.BignWithBash512 + " SupportedKeyClasses", bignKeyClasses);
 
-        put("Cipher.Belt", "by.bcrypto.bee2j.provider.BeltCipher");
-        put("Cipher.BeltECB", "by.bcrypto.bee2j.provider.BeltCipher$BeltECB");
-        put("Cipher.BeltCBC", "by.bcrypto.bee2j.provider.BeltCipher$BeltCBC");
-        put("Cipher.BeltCFB", "by.bcrypto.bee2j.provider.BeltCipher$BeltCFB");
-        put("Cipher.BeltCTR", "by.bcrypto.bee2j.provider.BeltCipher$BeltCTR");
-        put("Cipher.BeltMAC", "by.bcrypto.bee2j.provider.BeltCipher$BeltMAC");
-        put("Cipher.BeltDWP", "by.bcrypto.bee2j.provider.BeltCipher$BeltDWP");
+        put("Cipher.Belt", "by.bcrypto.bee2j.provider.cipher.BeltCipher");
+        put("Cipher.BeltECB", "by.bcrypto.bee2j.provider.cipher.BeltCipher$BeltECB");
+        put("Cipher.BeltCBC", "by.bcrypto.bee2j.provider.cipher.BeltCipher$BeltCBC");
+        put("Cipher.BeltCFB", "by.bcrypto.bee2j.provider.cipher.BeltCipher$BeltCFB");
+        put("Cipher.BeltCTR", "by.bcrypto.bee2j.provider.cipher.BeltCipher$BeltCTR");
+        put("Cipher.BeltMAC", "by.bcrypto.bee2j.provider.cipher.BeltCipher$BeltMAC");
+        put("Cipher.BeltDWP", "by.bcrypto.bee2j.provider.cipher.BeltCipher$BeltDWP");
 
         put("KeyPairGenerator.Bign", "by.bcrypto.bee2j.provider.BignKeyPairGenerator");
         put("Cipher.Bign", "by.bcrypto.bee2j.provider.by.BignCipherSpi");
